[ANDROSDK-1986] move api logic to network layer
